chore: Extract rename file hook
